+++ app/src/main/java/com/lft/spacex/ui/launches/LaunchesAdapter.java
@@ -49,6 +49,6 @@ public class LaunchesAdapter extends RecyclerView.Adapter<LaunchesHolder> {
     }
 
     public interface OnItemClickListener{
-        void onItemClick(long id);
+        void onItemClick(int id);
     }
 }
